Updated most Select value types to T
Select with a value of string is not helpful in cases where you have
integer ids. Began updating "Select" types to use generic type T that is
restricted to string | number.
Most types have been updated. Just need to brainstorm a way to deal with
the default value and when the value is stored in the DOM (in which case
it would have been coerced to a string).